requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
pydantic
faker
